+++ public/mage-app/src/ng1/factories/location.service.js
@@ -1,11 +1,10 @@
-var _ = require('underscore')
-  , $ = require('jquery');
+var _ = require('underscore');
 
 module.exports = LocationService;
 
-LocationService.$inject = ['$q', 'Location', 'UserService', 'LocalStorageService'];
+LocationService.$inject = ['$q', '$httpParamSerializer', 'Location', 'UserService', 'LocalStorageService'];
 
-function LocationService($q, Location, UserService, LocalStorageService) {
+function LocationService($q, $httpParamSerializer, Location, UserService, LocalStorageService) {
 
   // Specify times in milliseconds
   var colorBuckets = [{
@@ -46,7 +45,7 @@ function LocationService($q, Location, UserService, LocalStorageService) {
         UserService.getUser(userLocation.id).then(function(user) {
           if (user && user.iconUrl) {
             userLocation.locations[0].style = {
-              iconUrl: '/api/users/' + user.id + '/icon?' + $.param({access_token: LocalStorageService.getToken(), _dc: user.lastUpdated})
+              iconUrl: '/api/users/' + user.id + '/icon?' + $httpParamSerializer({access_token: LocalStorageService.getToken(), _dc: user.lastUpdated})
             };
           }