+++ public/mage-app/webpack.extra.js
@@ -10,9 +10,5 @@ module.exports = {
     }]
   },
   plugins: [
-    new webpack.ProvidePlugin({
-      'window.jQuery': 'jquery',
-      'jQuery': 'jquery'
-    })
   ]
 };